# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Hub — a workspace for distributed teams to manage projects, tasks, and people.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Member {
  id: number;
  name: string;
  email: string;
  /** Role label, e.g. "Designer", "Engineer", "PM" */
  role: string;
  /** Hex color used for the member's avatar fallback */
  avatarColor: string;
  createdAt: string;
}

export type MemberList = Member[];

export interface CreateMemberInput {
  name: string;
  email: string;
  role: string;
  /** @nullable */
  avatarColor?: string | null;
}

export interface UpdateMemberInput {
  name?: string;
  email?: string;
  role?: string;
  avatarColor?: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  archived: "archived",
} as const;

export interface Project {
  id: number;
  name: string;
  description: string;
  /** Hex color for project chip / accent */
  color: string;
  status: ProjectStatus;
  taskCount: number;
  completedCount: number;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ProjectList = Project[];

export interface CreateProjectInput {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  color?: string | null;
  status?: ProjectStatus;
  /** @nullable */
  dueDate?: string | null;
}

export interface UpdateProjectInput {
  name?: string;
  description?: string;
  color?: string;
  status?: ProjectStatus;
  /** @nullable */
  dueDate?: string | null;
}

export interface ProjectStats {
  projectId: number;
  totalTasks: number;
  todo: number;
  inProgress: number;
  inReview: number;
  done: number;
  /** 0-100 */
  completionPct: number;
  memberCount: number;
  overdueCount: number;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface Task {
  id: number;
  projectId: number;
  title: string;
  description: string;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
  /** Sort order within the task's status column. Lower values render first. */
  position: number;
  createdAt: string;
  updatedAt: string;
}

export type TaskList = Task[];

export interface CreateTaskInput {
  projectId: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status?: TaskStatus;
  priority?: TaskPriority;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface UpdateTaskInput {
  title?: string;
  description?: string;
  status?: TaskStatus;
  priority?: TaskPriority;
  /** @nullable */
  assigneeId?: number | null;
  /** @nullable */
  dueDate?: string | null;
  /** New sort position within the destination column. */
  position?: number;
}

export interface Comment {
  id: number;
  taskId: number;
  memberId: number;
  content: string;
  createdAt: string;
}

export type CommentList = Comment[];

export interface CreateCommentInput {
  memberId: number;
  content: string;
}

export interface Activity {
  id: number;
  /** Event type e.g. task_created, task_moved, task_completed, project_created, comment_added */
  type: string;
  description: string;
  memberId: number;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  taskId?: number | null;
  createdAt: string;
}

export type ActivityList = Activity[];

export type DashboardSummaryByStatus = {
  todo: number;
  inProgress: number;
  inReview: number;
  done: number;
};

export interface DashboardSummary {
  activeProjects: number;
  totalTasks: number;
  openTasks: number;
  completedThisWeek: number;
  overdueTasks: number;
  myOpenTasks: number;
  byStatus: DashboardSummaryByStatus;
}

export interface Workload {
  memberId: number;
  memberName: string;
  avatarColor: string;
  role: string;
  openTasks: number;
  todo: number;
  inProgress: number;
  inReview: number;
  done: number;
}

export type WorkloadList = Workload[];

/**
 * Not found
 */
export type NotFoundResponse = ErrorResponse;

export type ListProjectsParams = {
  status?: ListProjectsStatus;
};

export type ListProjectsStatus =
  (typeof ListProjectsStatus)[keyof typeof ListProjectsStatus];

export const ListProjectsStatus = {
  active: "active",
  archived: "archived",
} as const;

export type ListTasksParams = {
  projectId?: number;
  assigneeId?: number;
  status?: ListTasksStatus;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  todo: "todo",
  in_progress: "in_progress",
  in_review: "in_review",
  done: "done",
} as const;

export type GetRecentActivityParams = {
  limit?: number;
};