# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Hub — a workspace for distributed teams to manage projects, tasks, and people.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks with optional filters
      parameters:
        - in: query
          name: projectId
          required: false
          schema:
            type: integer
        - in: query
          name: assigneeId
          required: false
          schema:
            type: integer
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [todo, in_progress, in_review, done]
      responses:
        "200":
          description: A list of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskList"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by id
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Task with detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task (use to move between columns or edit details)
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Task deleted
...
components:
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TaskStatus:
      type: string
      enum: [todo, in_progress, in_review, done]
    TaskPriority:
      type: string
      enum: [low, medium, high, urgent]
    Task:
      type: object
      required:
        - id
        - projectId
        - title
        - description
        - status
        - priority
        - position
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        position:
          type: number
          description: Sort order within the task's status column. Lower values render first.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskList:
      type: array
      items:
        $ref: "#/components/schemas/Task"
    CreateTaskInput:
      type: object
      required: [projectId, title]
      properties:
        projectId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
    UpdateTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        assigneeId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        position:
          type: number
          description: New sort position within the destination column.
...